get Page to a fetchable state
