--- next-frontend/src/components/competitions/permissionsTestMessage.tsx
@@ -1,33 +1,0 @@
-"use client";
-
-import Link from "next/link";
-import { route } from "nextjs-routes";
-import { usePermissions } from "@/providers/PermissionProvider";
-import { components } from "@/types/openapi";
-import { Text, Link as ChakraLink } from "@chakra-ui/react";
-
-export default function PermissionsTestMessage({
-  competitionInfo,
-}: {
-  competitionInfo: components["schemas"]["CompetitionInfo"];
-}) {
-  const permissions = usePermissions();
-
-  return permissions?.canAdministerCompetition(competitionInfo.id) ? (
-    <Text>
-      You can administer this competition{" "}
-      <ChakraLink asChild variant="underline" colorPalette="teal">
-        <Link
-          href={route({
-            pathname: "/competitions/[competitionId]/admin",
-            query: { competitionId: competitionInfo.id },
-          })}
-        >
-          here
-        </Link>
-      </ChakraLink>
-    </Text>
-  ) : (
-    <p> This is the public page, you cannot administer this competition</p>
-  );
-}
