+++ next-frontend/src/lib/wca/data/wca.ts
@@ -1,0 +1,1 @@
+export const WCA_API_PAGINATION = 25;
